requirements.txt
discord.py-self
httpx
python-dotenv
